V2.3.5
Feels consistent and works well, but I don't like how it's doing it that much. For now I will let it be like that as long as it works but I would like to come back

Will be now working on data analysis python programming for summary stats and stuff like that lol.